undead-analysis
It works with the basic tests, should probably add some more edge cases.
There is something where the list will be reveresed except for the halt
case, so everything before halt is in reverse order. I have worked
around it by just removing the empty set, con it back, then reverse the
list so that it is in order